requirements.txt
beautifulsoup4==4.11.1
nest_asyncio==1.5.6
python-telegram-bot==21.7
Requests==2.32.3
